+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Sequence Java SDK changelog
 
-## 1.5 (2018????)
+## 1.5 (20180316)
 
 * Added `Feed`s. [More info](https://dashboard.seq.com/docs/feeds)
 * The `keys` field on `Account` and `Flavor` has been deprecated; the new field